move modules back to v1
