+++ sdk/python/kfp/dsl/artifact.py
@@ -21,7 +21,7 @@ from google.protobuf import json_format
 import yaml
 
 from kfp.pipeline_spec import pipeline_spec_pb2
-from kfp.v2.dsl import serialization_utils
+from kfp.dsl import serialization_utils
 
 _KFP_ARTIFACT_TITLE_PATTERN = 'kfp.{}'
 _KFP_ARTIFACT_ONTOLOGY_MODULE = 'kfp.v2.dsl.ontology_artifacts'
